graph_sgd: track number of nodes as a 'size_t' instead of 'int'
This squashes a number of compiler warnings and allows this struct to deal with
> 2³² - 1 nodes in future. Dealing with this number of nodes now is still not
possible due to locations where conversion back to `int` is necessary. In
future the source and target index arrays in this struct should probably be
converted to `size_t` too, but that was not done in this commit as it involves
more invasive changes.

+++ lib/neatogen/sgd.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -10,7 +12,7 @@ typedef struct term_sgd {
 } term_sgd;
 
 typedef struct graph_sgd {
-    int n; // number of nodes
+    size_t n; // number of nodes
     int *sources; // index of first edge in *targets for each node (length n+1)
     bool *pinneds; // whether a node is fixed or not
